Revert "Variable base scalar multiplication (#285)"
This reverts commit 2140c813bd4a07173dc419d4af31bbd0269ccf02.
