+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Anyhow` & `thiserror` for error handling support.
 - Serialisation methods for the crate public structures &
 `serde` support.
-- Add a `variable_base_scalar_mul` method using a variable base curve add gate.
+
 ### Removed
 - `failure` for error support since has been deprecated.
 
@@ -25,7 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 a `Scalar` and returns a constant & constrained witness `Variable`.
 - Update `add_witness_to_circuit_description` fn sig (#282, #284)
 - dusk-jubjub version updated to 0.3.6
-- `ecc::scalar_mul` now named fixed_base_scalar_mul
 
 ## [0.2.6] - 03-08-20
