+++ src/constraint_system/arithmetic.rs
@@ -65,8 +65,7 @@ impl StandardComposer {
         self.q_arith.push(Scalar::one());
         self.q_range.push(Scalar::zero());
         self.q_logic.push(Scalar::zero());
-        self.q_fixed_group_add.push(Scalar::zero());
-        self.q_variable_group_add.push(Scalar::zero());
+        self.q_ecc.push(Scalar::zero());
 
         self.public_inputs.push(pi);
 
@@ -142,8 +141,7 @@ impl StandardComposer {
 
         self.q_range.push(Scalar::zero());
         self.q_logic.push(Scalar::zero());
-        self.q_fixed_group_add.push(Scalar::zero());
-        self.q_variable_group_add.push(Scalar::zero());
+        self.q_ecc.push(Scalar::zero());
 
         self.public_inputs.push(pi);
