+++ src/constraint_system/boolean.rs
@@ -25,8 +25,7 @@ impl StandardComposer {
 
         self.q_range.push(Scalar::zero());
         self.q_logic.push(Scalar::zero());
-        self.q_fixed_group_add.push(Scalar::zero());
-        self.q_variable_group_add.push(Scalar::zero());
+        self.q_ecc.push(Scalar::zero());
 
         self.public_inputs.push(Scalar::zero());
