--- src/constraint_system/ecc/curve_addition/mod.rs
@@ -1,4 +1,0 @@
-/// A fixed base curve addition gate
-pub(crate) mod fixed_base_gate;
-/// A variable base curve addition gate
-pub(crate) mod variable_base_gate;
