--- src/constraint_system/ecc/scalar_mul/mod.rs
@@ -1,4 +1,0 @@
-/// Fixed base scalar multiplication gate
-pub mod fixed_base;
-/// Variable base scalar multiplication gate
-pub mod variable_base;
