--- src/proof_system/widget/ecc/curve_addition/mod.rs
@@ -1,5 +1,0 @@
-mod proverkey;
-mod verifierkey;
-
-pub use proverkey::ProverKey;
-pub use verifierkey::VerifierKey;
