--- src/proof_system/widget/ecc/scalar_mul/mod.rs
@@ -1,2 +1,0 @@
-pub mod fixed_base;
-pub mod variable_base;
