--- src/proof_system/widget/ecc/scalar_mul/variable_base/mod.rs
@@ -1,1 +1,0 @@
-
